[CHEF-3818] Remove the gemspec as this is a Rails application
The previous commit (88fbb11bf1) bump the Rails version in this gemspec
but missed bumping the Rails version in the Gemfile. As this is a Rails
application the gemspec is non-idiomatic and all dependencies should be
declared in the Gemfile (and locked in the Gemfile.lock). Removing this
gemspec now to save future confusion.